[step4] refactor: 패키지 리네이밍
