+++ src/main/java/racing/view/ResultView.java
@@ -1,4 +1,4 @@
-package racing.ui;
+package racing.view;
 
 import racing.dto.*;
