+++ src/test/java/racing/generator/StubValueGeneratorTest.java
@@ -1,4 +1,4 @@
-package racing.random;
+package racing.generator;
 
 import org.junit.jupiter.api.RepeatedTest;
